11. addede MovieDetail component.
This movie detail component renders when we click on the images in OurWork Section.
when we click we match the URL and then we render the movieDetail componenet accordingly. I MovieDetail we define the movie based on URL. We filter over the movies and we match url defined in movieState(data file) to our url defined in our browser.
We imported the MovieDetail in our App.js but the componenet only gets rendered when the url matches.